fix: createdBy in assetList
