+++ web/src/components/molecules/Asset/asset.type.ts
@@ -21,3 +21,5 @@ export type AssetFile = {
   path: string;
   size: number;
 };
+
+export type AssetNode = NonNullable<Asset>;
