+++ web/src/components/organisms/Asset/AssetList/assetSortType.type.ts
@@ -1,0 +1,14 @@
+import { AssetSortType as GQLSortType } from "@reearth-cms/gql/graphql-client-api";
+
+export type AssetSortType = "date" | "name" | "size";
+
+const enumTypeMapper: Partial<Record<GQLSortType, string>> = {
+  [GQLSortType.Date]: "date",
+  [GQLSortType.Name]: "name",
+  [GQLSortType.Size]: "size",
+};
+
+export function toGQLEnum(val?: AssetSortType) {
+  if (!val) return;
+  return (Object.keys(enumTypeMapper) as GQLSortType[]).find(k => enumTypeMapper[k] === val);
+}
