+++ web/src/components/organisms/Asset/convertAsset.ts
@@ -1,0 +1,17 @@
+import { Asset, AssetFile, PreviewType } from "@reearth-cms/components/molecules/Asset/asset.type";
+import { Asset as GQLAsset } from "@reearth-cms/gql/graphql-client-api";
+import { uuidToURL } from "@reearth-cms/utils/convert";
+
+export const convertAsset = (GQLAsset: GQLAsset | undefined): Asset => {
+  return {
+    id: GQLAsset?.id ?? "",
+    fileName: GQLAsset?.fileName ?? "",
+    createdAt: GQLAsset?.createdAt.toString() ?? "",
+    createdBy: GQLAsset?.createdBy?.name ?? "",
+    file: GQLAsset?.file as AssetFile,
+    previewType: GQLAsset?.previewType as PreviewType,
+    projectId: GQLAsset?.projectId ?? "",
+    size: GQLAsset?.size ?? 0,
+    url: GQLAsset?.uuid && GQLAsset?.fileName ? uuidToURL(GQLAsset?.uuid, GQLAsset?.fileName) : "",
+  };
+};
